R2 test moved to PostTaskSearchControllerCftTest class
