+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskmanagementapi/controllers/PostTaskSearchControllerTest.java
@@ -22,10 +22,8 @@ import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.SortingPar
 
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
@@ -34,7 +32,6 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
-import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.either;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.everyItem;
@@ -53,7 +50,6 @@ import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.Sea
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.SearchParameterKey.JURISDICTION;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.SearchParameterKey.LOCATION;
 import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.SearchParameterKey.STATE;
-import static uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.search.SearchParameterKey.WORK_TYPE;
 
 public class PostTaskSearchControllerTest extends SpringBootFunctionalBaseTest {
 
@@ -62,12 +58,7 @@ public class PostTaskSearchControllerTest extends SpringBootFunctionalBaseTest {
 
     private Headers authenticationHeaders;
 
-    private static final Map<String, String> TASK_TYPE_WORK_TYPE_MAP = new HashMap<>() {
-        {
-            put("arrangeOfflinePayment", "routine_work");
-            put("followUpOverdueReasonsForAppeal", "decision_making_work");
-        }
-    };
+
     private List<TestVariables> tasksCreated = new ArrayList<>();
     private String taskId1;
     private String taskId2;
@@ -708,172 +699,6 @@ public class PostTaskSearchControllerTest extends SpringBootFunctionalBaseTest {
             .body("total_records", greaterThanOrEqualTo(1));
     }
 
-    @Test
-    public void should_return_every_items_have_same_work_type_when_search_by_work_type() {
-        setupForR2();
-
-        String taskType = "followUpOverdueReasonsForAppeal";
-        TestVariables taskVariables = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables);
-        taskId1 = taskVariables.getTaskId();
-
-        initiateTask(taskId1, taskType, taskVariables);
-
-        SearchTaskRequest searchTaskRequest = new SearchTaskRequest(asList(
-            new SearchParameter(WORK_TYPE, SearchOperator.IN,
-                singletonList(TASK_TYPE_WORK_TYPE_MAP.get(taskType))),
-            new SearchParameter(CASE_ID, SearchOperator.IN,
-                singletonList(taskVariables.getCaseId()))
-        ), singletonList(new SortingParameter(SortField.DUE_DATE_SNAKE_CASE, SortOrder.DESCENDANT))
-        );
-        Response result = restApiActions.post(
-            ENDPOINT_BEING_TESTED,
-            searchTaskRequest,
-            authenticationHeaders
-        );
-
-        result.then().assertThat()
-            .statusCode(HttpStatus.OK.value())
-            .body("tasks.size()", greaterThanOrEqualTo(1))
-            .body("tasks.jurisdiction", everyItem(is("IA")))
-            .body("tasks.case_id", hasItem(taskVariables.getCaseId()))
-            .body("tasks.id", hasItem(taskId1))
-            .body("tasks.work_type", everyItem(is(TASK_TYPE_WORK_TYPE_MAP.get(taskType))))
-            .body("total_records", greaterThanOrEqualTo(1));
-    }
-
-    @Test
-    public void should_return_every_items_work_type_in_search_parameter_when_search_by_multiple_work_types() {
-        setupForR2();
-
-        //initiate first task
-        String taskType = "followUpOverdueReasonsForAppeal";
-        TestVariables taskVariables = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables);
-        taskId1 = taskVariables.getTaskId();
-
-        initiateTask(taskId1, taskType, taskVariables);
-
-        //initiate second task
-        taskType = "arrangeOfflinePayment";
-        TestVariables taskVariables2 = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables2);
-        taskId2 = taskVariables2.getTaskId();
-
-        initiateTask(taskId2, taskType, taskVariables2);
-
-        //search by all work types
-        SearchTaskRequest searchTaskRequest = new SearchTaskRequest(asList(
-            new SearchParameter(WORK_TYPE, SearchOperator.IN,
-                TASK_TYPE_WORK_TYPE_MAP.values().stream().collect(Collectors.toList())),
-            new SearchParameter(CASE_ID, SearchOperator.IN,
-                asList(taskVariables.getCaseId(), taskVariables2.getCaseId()))
-        ), singletonList(new SortingParameter(SortField.DUE_DATE_SNAKE_CASE, SortOrder.DESCENDANT))
-        );
-        Response result = restApiActions.post(
-            ENDPOINT_BEING_TESTED,
-            searchTaskRequest,
-            authenticationHeaders
-        );
-
-        result.then().assertThat()
-            .statusCode(HttpStatus.OK.value())
-            .body("tasks.size()", greaterThanOrEqualTo(2))
-            .body("tasks.jurisdiction", everyItem(is("IA")))
-            .body("tasks.case_id", contains(taskVariables2.getCaseId(), taskVariables.getCaseId()))
-            .body("tasks.id", contains(taskId2, taskId1))
-            .body("tasks.work_type", hasItems(
-                TASK_TYPE_WORK_TYPE_MAP.get("followUpOverdueReasonsForAppeal"),
-                TASK_TYPE_WORK_TYPE_MAP.get("arrangeOfflinePayment"))
-            )
-            .body("total_records", greaterThanOrEqualTo(2));
-    }
-
-    @Test
-    public void should_return_empty_list_when_search_by_work_type_not_exists() {
-        setupForR2();
-
-        //initiate first task
-        String taskType = "followUpOverdueReasonsForAppeal";
-        TestVariables taskVariables = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables);
-        taskId1 = taskVariables.getTaskId();
-
-        initiateTask(taskId1, taskType, taskVariables);
-
-        //initiate second task
-        taskType = "arrangeOfflinePayment";
-        TestVariables taskVariables2 = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables2);
-        taskId2 = taskVariables2.getTaskId();
-
-        initiateTask(taskId2, taskType, taskVariables2);
-
-        SearchTaskRequest searchTaskRequest = new SearchTaskRequest(asList(
-            new SearchParameter(WORK_TYPE, SearchOperator.IN, singletonList("aWorkType")),
-            new SearchParameter(CASE_ID, SearchOperator.IN,
-                asList(taskVariables.getCaseId(), taskVariables2.getCaseId()))
-        ), singletonList(new SortingParameter(SortField.DUE_DATE_SNAKE_CASE, SortOrder.DESCENDANT))
-        );
-
-        Response result = restApiActions.post(
-            ENDPOINT_BEING_TESTED,
-            searchTaskRequest,
-            authenticationHeaders
-        );
-
-        result.then().assertThat()
-            .statusCode(HttpStatus.OK.value())
-            .contentType(APPLICATION_JSON_VALUE)
-            .body("tasks.size()", equalTo(0));
-    }
-
-    @Test
-    public void should_return_every_items_work_type_in_search_parameter_when_search_by_work_types_and_case_id() {
-        setupForR2();
-
-        //initiate first task
-        String taskType = "followUpOverdueReasonsForAppeal";
-        TestVariables taskVariables = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables);
-        taskId1 = taskVariables.getTaskId();
-
-        initiateTask(taskId1, taskType, taskVariables);
-
-        //initiate second task
-        taskType = "arrangeOfflinePayment";
-        TestVariables taskVariables2 = common.setupTaskWithTaskIdAndRetrieveIds(taskType);
-        tasksCreated.add(taskVariables2);
-        taskId2 = taskVariables2.getTaskId();
-
-        initiateTask(taskId2, taskType, taskVariables2);
-
-        //search by all work types and caseIds
-        SearchTaskRequest searchTaskRequest = new SearchTaskRequest(asList(
-            new SearchParameter(WORK_TYPE, SearchOperator.IN,
-                TASK_TYPE_WORK_TYPE_MAP.values().stream().collect(Collectors.toList())),
-            new SearchParameter(CASE_ID, SearchOperator.IN,
-                asList(taskVariables.getCaseId(), taskVariables2.getCaseId()))
-        ));
-
-        Response result = restApiActions.post(
-            ENDPOINT_BEING_TESTED,
-            searchTaskRequest,
-            authenticationHeaders
-        );
-
-        result.then().assertThat()
-            .statusCode(HttpStatus.OK.value())
-            .body("tasks.size()", equalTo(2))
-            .body("tasks.jurisdiction", everyItem(is("IA")))
-            .body("tasks.case_id", hasItems(taskVariables.getCaseId(), taskVariables2.getCaseId()))
-            .body("tasks.id", hasItems(taskId1, taskId2))
-            .body("tasks.work_type", hasItems(
-                TASK_TYPE_WORK_TYPE_MAP.get("followUpOverdueReasonsForAppeal"),
-                TASK_TYPE_WORK_TYPE_MAP.get("arrangeOfflinePayment"))
-            )
-            .body("total_records", equalTo(2));
-    }
 
     private List<TestVariables> createMultipleTasks(String[] states) {
         for (String state : states) {
@@ -886,12 +711,6 @@ public class PostTaskSearchControllerTest extends SpringBootFunctionalBaseTest {
         }
 
         return tasksCreated;
-    }
-
-    private void setupForR2() {
-        authenticationHeaders = authorizationHeadersProvider
-            .getTribunalCaseworkerAAuthorization("Caseworker R2", "wa-ft-test-r2-");
-        common.setupOrganisationalRoleAssignment(authenticationHeaders);
     }
 
     private void initiateTask(String id, String taskType, TestVariables taskVariables) {
